[Bronze II] Title: 손익분기점, Time: 64 ms, Memory: 11452 KB -BaekjoonHub

+++ 백준/Bronze/1712. 손익분기점/손익분기점.java
@@ -9,9 +9,9 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         
         StringTokenizer st = new StringTokenizer(br.readLine(), " ");
-        long a = Long.parseLong(st.nextToken());
-        long b = Long.parseLong(st.nextToken());
-        long c = Long.parseLong(st.nextToken());
+        int a = Integer.parseInt(st.nextToken());
+        int b = Integer.parseInt(st.nextToken());
+        int c = Integer.parseInt(st.nextToken());
         
         if(b >= c) {
         	System.out.println(-1);
